Moved Areena classes to areena_extractors
